% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{Description of a data frame}
\usage{
contents(data)
}
\arguments{
\item{data}{a data frame.}
}
\value{
a list with 4 components:
\describe{
\item{dfname}{name of data frame}
\item{nrow}{number of rows}
\item{ncol}{number of columns}
\item{size}{size of the data frame in bytes}
\item{varinfo}{data frame of overall dataset characteristics}
}
}
\description{
\code{contents} provides describes the contents of a data
frame.
}
\details{
For each variable in a data frame, \code{contents} describes
the position, type, number of unique values, number of missing
values, and percent of missing values.
}
\examples{
df_info <- contents(happiness)
df_info
plot(df_info)
}
