% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edatools.R
\docType{package}
\name{edatools-package}
\alias{edatools-package}
\alias{edatools}
\title{Functions for exploring the contents of a data frame.}
\description{
edatools provides tools for exploring the variables in
a data frame.
}
