% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.contents}
\alias{plot.contents}
\title{Visualize a data frame}
\usage{
\method{plot}{contents}(x, ...)
}
\arguments{
\item{x}{an object of class \code{contents}.}

\item{...}{additional arguments (not currently used).}
}
\value{
a \code{ggplot2} graph
}
\description{
\code{plot.contents} visualizes the variables in a data frame.
}
\details{
For each variable, the plot displays
\itemize{
\item{type (\code{numeric},
\code{integer},
\code{factor},
\code{ordered factor},
\code{logical}, or \code{date})}
\item{percent of available (and missing) cases}
}
Variables are sorted by type and the total number of variables
and cases are printed in the caption.
}
\examples{
df_info <- contents(happiness)
plot(df_info)
}
\seealso{
See \link{contents}.
}
